moved SamplingParams
